refactor: useExpirationDate hook으로 분리
